Register selection with pointer
Instead of using a debounced function onSelectionChange, we only register selections on pointerUp events. This is much better for performance and is more in line with how most PDF readers behave.

As far as I can tell, the only reason onSelectionChange was used previously was due to the personal preference of the author of the original library.